requirements.txt
# Desktop GUI Framework
customtkinter==5.2.2
pillow==10.4.0

# Web Framework (for file server)
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.17

# Document Conversion - Updated for Python 3.12
pdf2docx==0.5.8
docx2pdf==0.1.8
python-docx==1.1.2
pypandoc==1.14

# PDF Handling
pypdf==5.1.0
reportlab==4.2.5

# Markdown Processing
markdown==3.7
weasyprint==62.3

# Async support
aiofiles==24.1.0
